docs: Create symlinks for NEWS/MAINTAINERS/CONTRIB
Some files needs to appear in the documentation and they are
needed in the 'docs' directory.

Creating symbolic links in the 'docs' directory for:

* docs/CONTRIBUTING -> CONTRIBUTING.md
* docs/NEWS.md -> NEWS
* docs/MAINTAINERS.md -> MAINTAINERS

This allows a legacy way of using those files from the root
directory of the project and for a new format and usage
by the mkdocs tool.

Signed-off-by: Pawel Sikora <pawel.sikora@intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>