igt_kms: Rename igt_pipe member enabled to valid
The boolean member enabled was added to struct igt_pipe by commit
b743593074ba ("lib/igt_kms: Add support for display with non-contiguous
pipes") to allow having array of pipes in struct igt_display to be
indexed by enum pipe.

The chosen terminology can be a bit confusing, because one could
interpret it as the pipe actively processing pixels.  Let's use "valid"
as an alternative name for that member to avoid the confusion.

Cc: Mohammed Khajapasha <mohammed.khajapasha@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Ravi Kumar V <ravi.kumar.vodapalli@intel.com>
Link: https://patch.msgid.link/20251029-rename-igt_pipe-dot-enabled-to-valid-v1-1-2aea1e009732@intel.com
Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>