lib/amdgpu: handle num_userq_slots compatibility for older driver
Ensure compatibility with older driver versions that may not
have the `num_userq_slots` field set in the `hw_ip_info` structure.
When this field is zero (indicating an older driver version),
set the default value of `available_rings` to 1 to ensure that at
least one ring is available for testing.

Signed-off-by: Jesse Zhang <Jesse.Zhang@amd.com>
Reviewed-by: Vitaly Prosyak <vitaly.prosyak@amd.com>