tests/dmabuf: Avoid soft lockups on slow platforms
Certain selftests, while basically correct, may fail on certain platforms.
E.g., igt@dmabuf@all-test@dma_fence_chain used to complete successfully,
but on slow machines it triggers soft lockup warnings which taint the
kernel.

Sometimes, like in the above mentioned case, it's not possible to fix a
root cause of the issue since it is not recognized as a bug.  To avoid
ever returning CI bug reports in such cases, allow selftests to be called
via user provided wrappers that take care of not triggering unavoidable
failures, e.g. by skipping specific selftests if some conditions are not
met, or watching their execution and acting upon certain conditions or
events.

With that in place, update the dmabuf test so it, as the first user of the
new feature, skips the dma_fence_chain selftest if a machine looks too
slow.  Since that's a hardware agnostic selftest, running it on a limited
subset of machines seems acceptable, especially when the soft lockups it
can trigger aren't recognized as bugs on the kernel side.

Signed-off-by: Janusz Krzysztofik <janusz.krzysztofik@linux.intel.com>
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>