tests/kms_vrr: Start virtual RR test above midpoint to avoid mode fallback
Starting from low_rr + step_size could cause the driver to fall back
to a lower physical mode due to proximity to the lower bound of the
VRR range. This results in an unintended clock change, which violates
the expectation of seamless VRR behavior.

To address this, the test now starts from above the midpoint of the
VRR range when multiple modes are available. This ensures the virtual
refresh rate remains safely within the active mode's VRR range and
avoids triggering a mode switch.

v2: Change logic to ternary operation. (Santhosh)

Signed-off-by: Naladala Ramanaidu <ramanaidu.naladala@intel.com>
Reviewed-by: Santhosh Reddy Guddati <santhosh.reddy.guddati@intel.com>