tests/intel/xe_peer2peer: gracefully handle ENOTSUP
Currently this blows up on the multi-dGPU setup we have in CI when
importing the VRAM from one device into the other via dma-buf, due the
P2P internals reporting this as unsupported on this machine. Trying to
manually white list the hostbridge device on the P2P side is also
unsuccessful since the read side test fails due to a data mismatch when
reading from the exported VRAM. There are also no hangs or errors in
dmesg although strangely the write side test passes. So it appears that
the direction of the copy is broken when reading the exported VRAM from
the importer side. Manually disabling/enabling iommu also make no
difference. And finally hacking the test to force same device
import/export makes the test pass, which should rule out any kind of
test defect. With that gracefully skip running the VRAM import subtests
if this is reported as unsupported by the kernel.

v2 (Kamil & Sanjay)
  - s/ENOSUP/ENOTSUP/ in commit title

Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/6433
Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/1173
Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Cc: Sanjay Yadav <sanjay.kumar.yadav@intel.com>
Reviewed-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>