tests/intel/xe_compute: Add eu-busy-10s test for EU utilization measurement
This test measures the utilization of available CCS instances based on the
number of slices reported in
/sys/class/drm/cardX/device/tileX/gtX/num_cslices. Compute workload is
scheduled on different CCS engines. The test runs workloads for fixed
duration using loop_kernel_duration. SR-IOV and CCS mode features are
mutually exclusive. This change updates the test to skip execution if
SR-IOV is enabled on the platform.

Signed-off-by: Nishit Sharma <nishit.sharma@intel.com>
Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Link: https://lore.kernel.org/r/20251112093510.15195-4-nishit.sharma@intel.com